requirements.txt
Django==2.0
Pillow==6.1.0
django-phonenumber-field==3.0.1
phonenumbers==8.10.17
selenium==3.141.0
requests==2.22.0
beautifulsoup4==4.8.0
bs4==0.0.1
django-ckeditor==5.7.1
django-truncate==0.1
httpie==1.0.3
django-hitcount==1.3.1
django-comments-xtd==2.4.0